#110 updating text and moving notification check
